Add Sign In, Sign Up with database

+++ new_app_spring_jpa/src/main/java/com/thieunv/data/repository/UserSignInRepository.java
@@ -8,7 +8,7 @@ import com.thieunv.data.model.UserSignIn;
  * Created by thieunv on 05/06/2017.
  */
 
-public interface UserSignInRepository extends JpaRepository<UserSignIn, Long> {
+public interface UserSignInRepository extends JpaRepository<UserSignIn, Integer> {
     UserSignIn findByEmail(String email);
     UserSignIn findByUserName(String userName);
 }
